Create index.html
is the main HTML file that represents the content of a webpage. It usually contains the layout, structure, and content of the webpage. In the context of the Flask application you provided, index.html is the template file that is rendered by the Flask server and served to the client's browser when they access the /execute route. The contents of this file define how the webpage will look and what elements it will contain.